User Positions table and model and view . Rbac setup

+++ controllers/actions/DeleteAction.php
@@ -1,8 +1,13 @@
 <?php
 namespace app\controllers\actions;
 
+use app\components\RbacComponent;
+
 class DeleteAction extends \yii\base\Action
 {
+    /**
+     * @var $rbac RbacComponent
+     */
     public $rbac;
     public $component;
     public $model;
@@ -15,10 +20,10 @@ class DeleteAction extends \yii\base\Action
      */
     public function run()
     {
-        //        if (!$this->rbac->canCreateActivity()){
-//            //throw new HttpException(403,'No access to create activity')
-//            return \Yii::$app->runAction('auth/signin');
-//        }
+        if (!$this->rbac->canCreateCatalog()){
+            //throw new HttpException(403,'No access to create activity')
+            return $this->controller->redirect(['/auth/login']);
+        }
         $component = \Yii::createObject(['class' => $this->component, 'nameClass' => $this->model]);
         if ($component->delete(\Yii::$app->request->queryParams)) {
             return $this->controller->redirect('index');
